+++ app/src/main/java/com/olderwold/reddit/MainViewModel.kt
@@ -3,8 +3,8 @@ package com.olderwold.reddit
 import androidx.lifecycle.ViewModel
 import androidx.paging.Pager
 import androidx.paging.PagingConfig
-import com.olderwold.reddit.domain.FeedItem
 import com.olderwold.reddit.feature.feed.FeedPagingSource
+import com.olderwold.reddit.feature.feed.domain.FeedItem
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
