+++ app/src/main/java/com/olderwold/reddit/feature/feed/FeedList.kt
@@ -23,7 +23,7 @@ import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
 import androidx.paging.compose.itemsIndexed
 import com.olderwold.reddit.MainViewModel
-import com.olderwold.reddit.domain.FeedItem
+import com.olderwold.reddit.feature.feed.domain.FeedItem
 import com.olderwold.reddit.feature.web.WebPage
 
 @Composable
