+++ app/src/main/java/com/olderwold/reddit/feature/feed/data/dto/Children.kt
@@ -1,7 +1,7 @@
-package com.olderwold.reddit.data.dto
+package com.olderwold.reddit.feature.feed.data.dto
 
 
-import com.olderwold.reddit.domain.FeedItem
+import com.olderwold.reddit.feature.feed.domain.FeedItem
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 import dagger.Reusable
