+++ app/src/main/java/com/olderwold/reddit/feature/feed/data/dto/MediaEmbed.kt
@@ -1,8 +1,8 @@
-package com.olderwold.reddit.data.dto
+package com.olderwold.reddit.feature.feed.data.dto
 
 
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
-class Variants(
+class MediaEmbed(
 )
