+++ app/src/main/java/com/olderwold/reddit/feature/feed/data/dto/Preview.kt
@@ -1,4 +1,4 @@
-package com.olderwold.reddit.data.dto
+package com.olderwold.reddit.feature.feed.data.dto
 
 
 import com.squareup.moshi.Json
