+++ app/src/main/java/com/olderwold/reddit/feature/feed/data/dto/ResizedStaticIcon.kt
@@ -1,11 +1,11 @@
-package com.olderwold.reddit.data.dto
+package com.olderwold.reddit.feature.feed.data.dto
 
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
-data class Source(
+data class ResizedStaticIcon(
     @Json(name = "height")
     val height: Int? = null,
     @Json(name = "url")
