+++ app/src/main/java/com/olderwold/reddit/feature/feed/domain/FeedItem.kt
@@ -1,4 +1,4 @@
-package com.olderwold.reddit.domain
+package com.olderwold.reddit.feature.feed.domain
 
 data class FeedItem(
     val id: String,
